feat: allow updating entities' IDs while keeping their names
This adds an additional step to the diff.Syncer's diff producer -
deleteDuplicates. Its purpose is to clean up the entities in the current state
that have duplicates with the same names in the target state before proceeding
with updates/creates and regular deletes. Without that additional step, it's
not possible to perform sync in which the IDs of entities are changed while
their names are kept between syncs.

An actual use-case that is the motivation for this change is Kong Ingress
Controller where we recently started assigning entities' IDs explicitly.
Without this change, existing setups break when KIC is upgraded from the
version not setting the IDs to the new one because of breaking the uniqueness
constraints of Services, Routes, and Consumers.

The added test cases represent the complete use case with all the affected
entities: Services, Routes, and Consumers. We change their IDs, keeping the
names and we expect the final state on the Kong side to be using the new IDs.